requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.4
nltk==3.9.1
numpy==2.0.2
openai==1.65.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
wordcloud==1.9.4
zipp==3.21.0
